+++ sql/999_version.sql
@@ -1,1 +1,1 @@
-SELECT set_version('0.6.6');
+SELECT set_version('0.6.7');