+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Changelog
+## [v0.6.7]
+### Added
+- Add get_queryables function to return a composite queryables json for either a single collection (text), a list of collections(text[]), or for the full repository (null::text).
+
+- Add missing_queryables(collection text, tablesample int) function to help identify if there are any properties in a collection without entries in the queryables table. The tablesample parameter is an int <=100 that is the approximate percentage of the collection to scan to look for missing queryables rather than reading every item.
+
+- Add missing_queryables(tablesample int) function that scans all collections using a sample of records to identify missing queryables.
+
+
 ## [v0.6.6]
 ### Added
 - Add support for array operators in CQL2 (a_equals, a_contains, a_contained_by, a_overlaps).